refactor: extract Animations to :common:android
It was decided that Preferences are per-profile, so this preference
needs to be profile-aware.

----

AnkiActivity needs to be moved to common for multimodule prep.
Therefore AnimUtils needs to move

This is the second step: exposing
'AnimationPreferences' in a profile-safe manner and moving AnimUtils.

Part of 20737
